更新BREW设置菜单
git-svn-id: svn://127.0.0.1@3374 3f3bbce5-bc85-8142-9eb1-77d090cfebd4

+++ main/modem/uim/uimdrv/src/uimtk.c
@@ -2553,7 +2553,7 @@ void uim_tk_send_cmd_to_ui
   ui_buf_ptr->proactive_cmd.hdr.task_ptr   = NULL;
   ui_buf_ptr->proactive_cmd.hdr.sigs       = 0;
 
-  ui_buf_ptr->proactive_cmd.num_bytes = (byte)rsp_ptr->cmd_rsp_size;
+  ui_buf_ptr->proactive_cmd.num_bytes = (uint16)rsp_ptr->cmd_rsp_size;
   (void) memcpy ( ui_buf_ptr->proactive_cmd.cmd_data,
                   (void *) rsp_ptr->rsp.data, rsp_ptr->cmd_rsp_size );
